2010-07-24  Pavel Feldman  <pfeldman@chromium.org>
        Not reviewed: chromium build fix.

        * src/ApplicationCacheHost.cpp:
        (WebCore::ApplicationCacheHost::fillResourceList):


Canonical link: https://commits.webkit.org/54841@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@64009 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebKit/chromium/src/ApplicationCacheHost.cpp
@@ -228,7 +228,7 @@ void ApplicationCacheHost::fillResourceList(ResourceInfoList* resources)
 
     WebKit::WebVector<WebKit::WebApplicationCacheHost::ResourceInfo> webResources;
     m_internal->m_outerHost->getResourceList(&webResources);
-    for (int i = 0; i < webResources.size(); ++i) {
+    for (size_t i = 0; i < webResources.size(); ++i) {
         resources->append(ResourceInfo(
             webResources[i].url, webResources[i].isMaster, webResources[i].isManifest, webResources[i].isFallback,
             webResources[i].isForeign, webResources[i].isExplicit, webResources[i].size));
